Move the outgoing body types to body.rs
